+++ upgrades/blocks.go
@@ -1,0 +1,4 @@
+package upgrades
+
+const UpgradeBlock0 = 5760
+const UpgradeBlock1 = 250000