+++ CHANGELOG.md
@@ -11,6 +11,7 @@ BUG FIXES
 
 - [api] Fix block proposer issue
 - [api] Fix 0x api error
+- [core] Fix "Stake is too low" issue with new custom coins
 
 ## 1.0.0