+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## 1.0.2
+
+IMPROVEMENT
+
+- [config] Add new seed nodes
+- [api] Add /missed_blocks endpoint
+
+BUG FIXES
+
+- [api] Fix block proposer issue
+- [api] Fix 0x api error
+- [core] Fix "Stake is too low" issue with new custom coins
+- [core] Fix redeem check invariants
+
 ## 1.0.0
 
 BUG FIXES